requirements.txt
Django==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.3
Pillow==10.2.0
psycopg2-binary==2.9.9
cryptography==42.0.5
requests==2.31.0
reportlab==4.0.9
openpyxl==3.1.2
google-auth==2.25.2
google-auth-httplib2==0.2.0

